Test customer colour (#34)
* Test CustomerColour not finished yet

* Change ColourDao Function row 128

* Implemented Function to change and delete colourMixture and colour.

* Fixed DeleteColour and Update CustomerColour

* Fix CustomerColour and get All from DB

* Fix getallappointments

+++ src/main/java/data_loader/data_access_object/ColourMixtureDao.java
@@ -51,11 +51,11 @@ public class ColourMixtureDao {
         }
         return colourMixture;
     }
-
     public static ColourMixture getColourMixtureByColourMixtureId(String colourMixtureId){
         ColourMixture colourMixture = null;
         try {
-            preparedStmt = con.prepareStatement("SELECT * FROM OPTKOS.COLOURMIXTURE WHERE COLOURMIXTUREID=?");
+            preparedStmt = con.prepareStatement("SELECT * FROM OPTKOS.COLOURMIXTURE cm, OPTKOS.COLOUR c" +
+                    "WHERE cm.COLOURID = c.COLOURID AND COLOURMIXTUREID=?");
             preparedStmt.setString(1, colourMixtureId);
             try(ResultSet rs = preparedStmt.executeQuery()){
                 while(rs.next()){
@@ -74,10 +74,10 @@ public class ColourMixtureDao {
             preparedStmt= con.prepareStatement("INSERT INTO OPTKOS.COLOURMIXTURE (MIXINGRATIO, COLOURID," +
                     " COLOURMIXTUREID, CUSTOMERID) VALUES(?,?,?,?)");
             preparedStmt.setInt(1, colourMixture.getMixingRatio());
-            preparedStmt.setString(2, colourMixture.getColourId());
+            preparedStmt.setString(2, colourMixture.getColour().getColourId());
             preparedStmt.setString(3, colourMixture.getColourMixtureId());
             preparedStmt.setString(4, colourMixture.getCustomerId());
-            preparedStmt.executeUpdate();
+            preparedStmt.execute();
             preparedStmt.close();
         } catch (SQLException e) {
             e.printStackTrace();
@@ -85,37 +85,40 @@ public class ColourMixtureDao {
     }
 
     public static void bulkCreate(List<ColourMixture> cmList){
-        StringBuilder query = new StringBuilder();
-        query.append("INSERT INTO OPTKOS.COLOURMIXTURE (MIXINGRATIO, COLOURID, COLOURMIXTUREID, CUSTOMERID) VALUES");
-        for (int i = 0; i<cmList.size(); i++){
-            query.append("(?,?,?,?),");
-        }
-        query.deleteCharAt(query.length()-1);
-        try {
-            preparedStmt = con.prepareStatement(query.toString());
-            int index = 1;
-            for (ColourMixture cm :
-                    cmList) {
-                preparedStmt.setInt(index++, cm.getMixingRatio());
-                preparedStmt.setString(index++, cm.getColourId());
-                preparedStmt.setString(index++, cm.getColourMixtureId());
-                preparedStmt.setString(index++, cm.getCustomerId());
+        if(cmList.size()>0) {
+            StringBuilder query = new StringBuilder();
+            query.append("INSERT INTO OPTKOS.COLOURMIXTURE (MIXINGRATIO, COLOURID, COLOURMIXTUREID, CUSTOMERID) VALUES");
+            for (int i = 0; i < cmList.size(); i++) {
+                query.append("(?,?,?,?),");
             }
-            preparedStmt.executeUpdate();
-            preparedStmt.close();
-        } catch (SQLException e) {
-            System.err.println("ERROR while bulkcreating colourMixtures");
-            e.printStackTrace();
+            query.deleteCharAt(query.length() - 1);
+            try {
+                preparedStmt = con.prepareStatement(query.toString());
+                int index = 1;
+                for (ColourMixture cm :
+                        cmList) {
+                    preparedStmt.setInt(index++, cm.getMixingRatio());
+                    preparedStmt.setString(index++, cm.getColour().getColourId());
+                    preparedStmt.setString(index++, cm.getColourMixtureId());
+                    preparedStmt.setString(index++, cm.getCustomerId());
+                }
+                preparedStmt.execute();
+                preparedStmt.close();
+            } catch (SQLException e) {
+                System.err.println("ERROR while bulkcreating colourMixtures");
+                e.printStackTrace();
+            }
         }
-
     }
 
-    public static void deleteColourMixtureByColourMixtureId(String colourMixtureId){
+    public static void deleteColourMixtureByColourMixtureId(String colourMixtureId, String colourId){
         try {
-            preparedStmt = con.prepareStatement("DELETE FROM OPTKOS.COLOURMIXTURE WHERE COLOURMIXTUREID=?");
-            preparedStmt.setString(1, colourMixtureId);
+            preparedStmt = con.prepareStatement("DELETE FROM OPTKOS.COLOURMIXTURE WHERE COLOURID=?");
+            preparedStmt.setString(1, colourId);
             preparedStmt.executeUpdate();
             preparedStmt.close();
+
+            ColourDao.deleteColourByColourId(colourId);
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -133,25 +136,30 @@ public class ColourMixtureDao {
             e.printStackTrace();
         }
     }
+    public static void deleteAllColourMixturesbyCustomerId(String customerID){
 
-    @SuppressWarnings("Duplicates")
-    public static void bulkDeleteById(List<String> cmIdList){
-        StringBuilder sbQuery = new StringBuilder();
-        sbQuery.append("DELETE FROM OPTKOS.COLOURMIXTURE WHERE COLOURMIXTUREID IN (");
-        for (String s :
-                cmIdList) {
-            sbQuery.append("?,");
+        try {
+            preparedStmt = con.prepareStatement("DELETE FROM OPTKOS.COLOURMIXTURE cm JOIN OPTKOS.COLOUR c ON c.COLOURID = cm.COLourid WHERE CUSTOMERID=?");
+            preparedStmt.setString(1, customerID);
+            preparedStmt.executeUpdate();
+            preparedStmt.close();
+        } catch (SQLException e) {
+            e.printStackTrace();
         }
-        sbQuery.deleteCharAt(sbQuery.length()-1);
+    }
+    @SuppressWarnings("Duplicates")
+    public static void bulkDeleteById(List<String> colorIds){
+        StringBuilder sbQuery = new StringBuilder();
+        sbQuery.append("DELETE FROM OPTKOS.COLOURMIXTURE WHERE COLOURID IN (");
+        for (String s :
+                colorIds) {
+            sbQuery.append("'" + s + "', ");
+        }
+        sbQuery.deleteCharAt(sbQuery.length()-2);
         sbQuery.append(")");
 
         try {
             preparedStmt = con.prepareStatement(sbQuery.toString());
-            int index = 1;
-            for (String s :
-                    cmIdList) {
-                preparedStmt.setString(index++, s);
-            }
 
             preparedStmt.executeUpdate();
             preparedStmt.close();
@@ -162,19 +170,18 @@ public class ColourMixtureDao {
 
     public static void addNewMixtures(List<ColourMixture> colourMixturesList){
         ArrayList<String> colorIds = new ArrayList<>();
-        ArrayList<String> cmIds = new ArrayList<>();
         ArrayList<Colour> colors = new ArrayList<>();
 
         /*Get IDs*/
         for(ColourMixture c: colourMixturesList) {
             colorIds.add(c.getColour().getColourId());
             colors.add(c.getColour());
-            cmIds.add(c.getColourMixtureId());
         }
 
         /*Delete all*/
+        bulkDeleteById(colorIds);
         ColourDao.bulkDeleteById(colorIds);
-        bulkDeleteById(cmIds);
+
 
         /*Create colours*/
         ColourDao.bulkcreate(colors);
